move wasm impl to signer crate
