+++ sdk/signer/Cargo.toml
@@ -17,6 +17,9 @@ solana-pubkey = { workspace = true }
 solana-signature = { workspace = true }
 solana-transaction-error = { workspace = true }
 
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+wasm-bindgen = { workspace = true }
+
 [dev-dependencies]
 serde_json = { workspace = true }
 solana-signer = { path = ".", features = ["dev-context-only-utils"] }
